create component to Modal
